Make the $repo_root for the container be created before running docker
Before:

The $repo_root for inside the container was being created by the docker.

Why change:

When docker creates the $repo_root directory, it uses the `root` user.
We want it to use the user who is running the process instead.

Now:

The "$container_root$repo_root" directory that is mounted into the
container is created before docker is run by the user running docker so
the user will own the directory and be able to delete it afterward.

Fixes: cockroachdb/dev-inf#86