avoid unnecessary USING clause in ALTER COLUMN TYPE
When altering a column's type, the USING clause is only necessary if the
underlying data type is changing. Previously the clause was always
included, which causes CockroachDB v20.2 to reject type conversions that
are actually safe. I added logic to detect whether the USING clause is
necessary.

Note that this is not an issue in 3.0.x, which already has this logic
built in.